np exp requires float so ensured s_abilities returns as dtype float

+++ high_dim_bt/src/high_dim_bt/nodes/data_engineering.py
@@ -133,4 +133,4 @@ def get_starting_abilities(
     abilities = data[[winner_pts, loser_pts]].values.ravel()[pos]
 
     # any nan's relpaced with 0
-    return np.nan_to_num(abilities, nan=0)
+    return np.nan_to_num(abilities, nan=0).astype(float)
